scripts/docs/requirements.txt
# Dependencies for report PDF generation (generate-report-pdf.py)
pyyaml>=5.4
